修复BUG,完善细节

+++ FourNews/FourNews/Class/AV/Tool_Funtion/FNAVGetAVNewsList.m
@@ -14,7 +14,7 @@
 
 + (void)getAVNewsListWithTid:(NSString *)tid :(NSInteger)pageCount :(void (^)(NSArray *))complete
 {
-    NSString *urlStr = [NSString stringWithFormat:@"http://c.m.163.com/nc/video/Tlist/%@/%ld0-10.html",tid,pageCount];
+    NSString *urlStr = [NSString stringWithFormat:@"http://c.m.163.com/nc/video/Tlist/%@/%ld0-10.html",tid,(long)pageCount];
     [FNNetWorking GET:urlStr parameters:nil progress:^(NSProgress *progress) {
     } success:^(id responseObject, NSURLSessionDataTask *task) {
         NSArray *items = [FNAVListItem mj_objectArrayWithKeyValuesArray:responseObject[tid]];
